Install the new ElimByName in place of ByNameClosures
... and drop the old ElimByName.
